# openapi/commons.yaml
openapi: 3.0.3
info:
  title: Commons Visualization & Query API
  version: '0.1.0'
  description: |
    Read-only, privacy-preserving access to aggregate regeneration metrics,
    time-series data, ecosystem health indices, and public audit trails.
    Transparency without surveillance.
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/commons/audit-trail:
    get:
      tags: [audit]
      summary: Public audit events (redacted)
      description: >-
        Append-only public audit trail with sensitive coordinates and identifiers
        redacted. Enables transparency and verification.
      parameters:
        - in: query
          name: since
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Audit events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditTrailResponse'
components:
  schemas:
...
    AuditEvent:
      type: object
      properties:
        eventId:
          type: string
        timestamp:
          type: string
          format: date-time
        eventType:
          type: string
          enum: [credential_issued, credential_revoked, activation_triggered, policy_published]
        subject:
          type: string
          description: Redacted subject identifier
        action:
          type: string
        hash:
          type: string
          description: Merkle tree hash for integrity verification
        redactions:
          type: array
          items:
            type: string
    AuditTrailResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        hasMore:
          type: boolean
        cursor:
          type: string
        integrityRoot:
          type: string
          description: Latest Merkle root for audit trail verification
...